% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_missing_or_empty_character.R
\name{is_missing_or_empty_character}
\alias{is_missing_or_empty_character}
\alias{all_are_missing_or_empty_character}
\title{Are all missing or empty character}
\usage{
is_missing_or_empty_character(x)

all_are_missing_or_empty_character(x)
}
\arguments{
\item{x}{character vector}
}
\value{
TRUE or FALSE
}
\description{
Are all missing or empty character
}
\examples{
x <- c('a', 'b', NA_character_, 'd', '')
all_are_missing_or_empty_character(x)
}
