% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_identical_values.R
\name{all_have_value}
\alias{all_have_value}
\title{All have value}
\usage{
all_have_value(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{scalar}
}
\value{
logical(1)
}
\description{
All have value
}
\examples{
require(magrittr)
c(1,1,1,1) \%>\% all_have_value(1)
}
