% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_svalues_available.R
\name{all_svalues_available}
\alias{all_svalues_available}
\title{Are all svalues available?}
\usage{
all_svalues_available(object, svar)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{svar}{character(1)}
}
\value{
logical(1)
}
\description{
Are all svalues available?
}
