% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{cdraw}
\alias{cdraw}
\title{Convenient grid.draw}
\usage{
cdraw(x)
}
\arguments{
\item{x}{grid object}
}
\description{
Convenient grid.draw
}
\examples{
my_circle <- grid::circleGrob(x = 0.5, y = 0.5, r = 0.5)
cdraw(my_circle)
}
