% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated.R
\name{cduplicated}
\alias{cduplicated}
\title{Convenient (two way) duplicated}
\usage{
cduplicated(x)
}
\arguments{
\item{x}{vector}
}
\value{
logical vector
}
\description{
Convenient (two way) duplicated
}
\examples{
require(magrittr)
c(1,2,3,4,5,2) \%>\% cduplicated()
}
