% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_max.R
\name{cequals}
\alias{cequals}
\title{Convenient equals operator}
\usage{
cequals(x, y)
}
\arguments{
\item{x}{numeric vector or scalar}

\item{y}{numeric scalar}
}
\value{
logical vector
}
\description{
Performs x == y, but returns FALSE rather than NA for NA elements of x.
}
\examples{
x <- c(A=1,B=3,C=2,D=3, E=NA)
y <- 3
equals(x, y)
}
