% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_path.R
\name{cfile.path}
\alias{cfile.path}
\title{convenient interface to file.path}
\usage{
cfile.path(...)
}
\arguments{
\item{...}{sent to file.path}
}
\description{
removes redundant slashes before running file.path
}
\examples{
cfile.path('dir/', 'subdir')
}
