% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_operations.R
\name{characterify}
\alias{characterify}
\title{Characterify a dataframe}
\usage{
characterify(x, char_length = 40)
}
\arguments{
\item{x}{dataframe}

\item{char_length}{allowed length of strings}
}
\value{
compactified dataframe
}
\description{
Transform numeric variables into character variables in scientific notation
Limit length of character variables
}
\examples{
characterify(data.frame(a = 1.09736, b = 4.902903))
}
