% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{cmessage}
\alias{cmessage}
\title{Conveniently message}
\usage{
cmessage(format_string, ...)
}
\arguments{
\item{format_string}{sprintf format string}

\item{...}{additional arguments  passed to sprintf}
}
\description{
Print message to screen with sprintf syntax
}
\examples{
cmessage('\\t\%s\\t\%s', 'Hi', 'there')
}
