% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{cmessage_df}
\alias{cmessage_df}
\title{Conveniently message dataframe}
\usage{
cmessage_df(format_string, x)
}
\arguments{
\item{format_string}{sprintf style format string}

\item{x}{data.frame}
}
\description{
Conveniently message dataframe using sprintf syntax.
Use place holder '%s' for data.frame.
}
\examples{
x <- data.frame(feature_id = c('F001', 'F002'), symbol = c('FEAT1', 'FEAT2'))
cmessage_df('\t\%s', x)

x <- c(rep('PASS', 25), rep('FAIL', 25))
cmessage_df(format_string = '\%s', table(x))
}
