% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonify_strings.R
\name{commonify_two_strings}
\alias{commonify_two_strings}
\alias{commonify_strings}
\title{Commonify strings from their heads}
\usage{
commonify_two_strings(x, y)

commonify_strings(x)
}
\arguments{
\item{x}{string}

\item{y}{string}
}
\description{
Commonify strings from their heads
}
\examples{
require(magrittr)
# Example 1: tail slightly different
x <- "Golgi apparatus membrane protein TVP23 homolog B"
y <- "Golgi apparatus membrane protein TVP23 homolog C"
z <- "Golgi apparatus membrane protein TVP23 homolog D"
commonify_two_strings(x,y)
commonify_strings(c(x,y,z))

# Example 2: tail different
x <- 'Histone H2B type 1-K'
y <- 'Histone H2B type 1-C/E/F/G/I'
commonify_two_strings(x,y)

# Example 3: all common
x <- 'Retrotransposon Gag-like protein 8B'
y <- 'Retrotransposon Gag-like protein 8B'
commonify_two_strings(x,y)

}
\references{
https://stackoverflow.com/questions/28261825/longest-common-substring-in-r-finding-non-contiguous-matches-between-the-two-str
}
