% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_operations.R
\name{dedupe_varnames}
\alias{dedupe_varnames}
\title{Dedupe varnames}
\usage{
dedupe_varnames(x)
}
\arguments{
\item{x}{list}
}
\description{
Dedupe varnames
}
\examples{
require(magrittr)
x <- cbind(data.frame(a=1:3, b=1:3, c=1:3), data.frame(a=4:6))
x
x \%>\% autonomics.support::dedupe_varnames()
}
