% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_even.R
\name{evenify_upwards}
\alias{evenify_upwards}
\title{Evenify upwards}
\usage{
evenify_upwards(x)
}
\arguments{
\item{x}{integer}
}
\value{
integer
}
\description{
Evenify upwards
}
\examples{
evenify_upwards(3)
}
