% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_max.R
\name{extract_max}
\alias{extract_max}
\title{Extract portion of vector that contains maximum values}
\usage{
extract_max(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Extract portion of vector that contains maximum values
}
\examples{
require(magrittr) 
x <- c(A=1,B=3,C=2,D=3, E=NA)
x \%>\% autonomics.support::extract_max()
}
