% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{factorify}
\alias{factorify}
\title{Create factor with levels in order of appearance}
\usage{
factorify(avector, reverse = FALSE)
}
\arguments{
\item{avector}{An atomic vector.}

\item{reverse}{A logical value. Should factor levels be in reverse order of
appearance.}
}
\value{
A factor.
}
\description{
Creates a factor from a vector, where the levels are in (possibly reverse)
order of appearance in the vector, rather than being alphabetically sorted.
}
