% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_tails.R
\name{get_unique_tails}
\alias{get_unique_tails}
\title{Get unique unique tails}
\usage{
get_unique_tails(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Get unique unique tails
}
\examples{
x <- c("E_1", "E_2", "E_3")
x \%>\% get_unique_tail()
}
