% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_even.R
\name{has_even_length}
\alias{has_even_length}
\alias{has_odd_length}
\title{Has even/odd length?}
\usage{
has_even_length(x)

has_odd_length(x)
}
\arguments{
\item{x}{vector}
}
\value{
logical
}
\description{
Has even/odd length?
}
\examples{
has_even_length(1:2)
has_odd_length(1:2)
has_even_length(1:3)
has_even_length(1:3)
}
