% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_identical_values.R
\name{has_identical_values}
\alias{has_identical_values}
\title{All elements of vector are identical}
\usage{
has_identical_values(x)
}
\arguments{
\item{x}{vector}
}
\value{
TRUE or FALSE
}
\description{
All elements of vector are identical
}
\examples{
x <- c(2,2,1,2)
has_identical_values(x)
}
