% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_greater_than.R
\name{is_greater_than}
\alias{is_greater_than}
\alias{is_weakly_greater_than}
\alias{is_less_than}
\alias{is_weakly_less_than}
\title{NA-accepting versions of comparator operators}
\usage{
is_greater_than(a, b)

is_weakly_greater_than(a, b)

is_less_than(a, b)

is_weakly_less_than(a, b)
}
\arguments{
\item{a}{numeric scalar, vector, or matrix}

\item{b}{numeric scalar, vector, or matrix}
}
\value{
logical scalar, vector, or matrix
}
\description{
NA-accepting versions of comparator operators
}
