% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_max.R
\name{is_max}
\alias{is_max}
\title{Is maximal}
\usage{
is_max(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
logical vector
}
\description{
Is maximal
}
\examples{
x <- c(A=1,B=3,C=2,D=3, E=NA)
is_max(x)
}
