% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_outlier.R
\name{is_outlier}
\alias{is_outlier}
\title{Identify outliers in vector}
\usage{
is_outlier(x)
}
\arguments{
\item{x}{numeric}
}
\value{
logical
}
\description{
Identify outliers in vector
}
\examples{
require(magrittr)
c(1,2,3,4,5,6, 200) \%>\% autonomics.support::is_outlier()
}
