% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_weakly_greater_than.R
\name{is_weakly_greater_than}
\alias{is_weakly_greater_than}
\title{NA-accepting is_weakly_greater_then}
\usage{
is_weakly_greater_than(a, b)
}
\arguments{
\item{a}{numeric scalar, vector, or matrix}

\item{b}{numeric scalar, vector, or matrix}
}
\value{
logical scalar, vector, or matrix
}
\description{
NA-accepting is_weakly_greater_then
}
