% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_operations.R
\name{left_join_keeping_rownames}
\alias{left_join_keeping_rownames}
\title{Left join while preserving rownames}
\usage{
left_join_keeping_rownames(x, y, by, ...)
}
\arguments{
\item{x}{first object}

\item{y}{second object}

\item{by}{by}

\item{...}{passed to dplyr::left_join}
}
\value{
merged dataframe
}
\description{
Convenient left_join function that preserves row names
and avoids warning messages when by values are of different type.
}
\examples{
x <- data.frame(organism = c('Homo sapiens', 'Mus musculus'), 
        commonname   = c('human', 'mouse'), row.names = c('Hs', 'Mm'))
y <- data.frame(organism = c('Homo sapiens', 'Mus musculus'),
        chromosomes = c(23, 20), row.names = c('Hs', 'Mm'))
left_join_keeping_rownames(x, y, by = 'organism')
}
\seealso{
dplyr::left_join
}
