% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load data}
\usage{
load_data(x, package = NULL)
}
\arguments{
\item{x}{character vector}

\item{package}{package name (character)}
}
\value{
object
}
\description{
Programmatic equivalent of data(.)
}
\examples{
if (require(autonomics.data)){
   load_data('car90', package = 'rpart')
}
}
\references{
http://stackoverflow.com/questions/30951204/load-dataset-from-r-package-using-data-assign-it-directly-to-a-variable
}
