% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_operations.R
\name{matrixify}
\alias{matrixify}
\title{Matrixify datatable or dataframe}
\usage{
matrixify(df)
}
\arguments{
\item{df}{dataframe or datatable (first column contains rownames)}
}
\value{
matrix
}
\description{
Matrixify datatable or dataframe
}
