% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmax.R
\name{nmax}
\alias{nmax}
\alias{nmin}
\title{Return nth max (min) value in vector}
\usage{
nmax(x, n)

nmin(x, n)
}
\arguments{
\item{x}{numeric vector}

\item{n}{integer}
}
\value{
value
}
\description{
Orders a vector and returns n'th ordered value.
When vector length is smaller than n, returns last value.
}
