% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_operations.R
\name{nullify}
\alias{nullify}
\title{Nullify variables in df}
\usage{
nullify(df, vars)
}
\arguments{
\item{df}{data.frame}

\item{vars}{character vector with variable names}
}
\value{
dataframe
}
\description{
Nullify variables in df
}
\examples{
require(magrittr)
data.frame(a = 1:3, b = 4:6, c = 7:9, d = 10:12) \%>\% 
autonomics.support::nullify(c('a', 'b'))
}
