% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_integer_vector.R
\name{pad_integer_vector}
\alias{pad_integer_vector}
\title{Pad integer vector}
\usage{
pad_integer_vector(x, pad = "0")
}
\arguments{
\item{x}{integer vector}

\item{pad}{padding character}
}
\value{
character vector
}
\description{
Pad integer vector
}
