% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_available.R
\name{pct_available}
\alias{pct_available}
\title{Percentage of available values}
\usage{
pct_available(x)
}
\arguments{
\item{x}{vector}
}
\value{
rounded percentage
}
\description{
Percentage of available values
}
\examples{
x <- c(NA, 2, 3, NA, 5, 6)
pct_available(x)
}
