% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_available.R
\name{pct_true}
\alias{pct_true}
\title{Percentage of TRUE value}
\usage{
pct_true(x)
}
\arguments{
\item{x}{logical vector}
}
\value{
rounded percentage
}
\description{
Percentage of TRUE value
}
\examples{
x <- c(TRUE, FALSE, TRUE, TRUE)
pct_true(x)
}
