% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print2pdf}
\alias{print2pdf}
\title{Print graphics object to pdf}
\usage{
print2pdf(object, file, width = 7, height = 7, ...)
}
\arguments{
\item{object}{graphics object}

\item{file}{file path to write to}

\item{width}{figure width}

\item{height}{figure height}

\item{...}{additional arguments passed to the pdf() function}
}
\description{
Print graphics object to pdf
}
