% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print2txt}
\alias{print2txt}
\title{Print dataframe to txt}
\usage{
print2txt(a_df, file = "", quote = FALSE, sep = "\\t", row.names = FALSE, ...)
}
\arguments{
\item{a_df}{A \code{dta.frame}, or similar.}

\item{file}{Either a character string naming a file or a connection open for 
writing. Passed to \code{\link[utils]{write.table}}.}

\item{quote}{Passed to \code{\link[utils]{write.table}}.}

\item{sep}{Passed to \code{\link[utils]{write.table}}.}

\item{row.names}{Passed to \code{\link[utils]{write.table}}.}

\item{...}{Passed to \code{\link[utils]{write.table}}.}
}
\description{
Writes a data frame-like object to a tab delimited text file.
}
\examples{
require(magrittr)
my_file <- tempfile() \%T>\% message()
data.frame(a=1, b=2, c=3) \%>\% print2txt(file = my_file)
}
\seealso{
\code{\link[data.table]{fwrite}}
}
