% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_columns.R
\name{pull_columns}
\alias{pull_columns}
\title{Pull columns in a dataframe to the front}
\usage{
pull_columns(df, first_cols)
}
\arguments{
\item{df}{dataframe}

\item{selected_cols}{}
}
\value{
dataframe with re-ordered columns
}
\description{
Pull columns in a dataframe to the front
}
\examples{
require(magrittr)
df <- data.frame(
   symbol = c('A1BG', 'A2M'), 
   id     = c('1',    '2'),
   name   = c('alpha-1-B glycoprotein', 'alpha-2-macroglobulin'), 
   type   = c('proteincoding', 'proteincoding'))
first_cols <- c('id', 'symbol') 
df \%>\% autonomics.support::pull_columns(first_cols)
}
