% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_empty_vars.R
\name{rm_empty_vars}
\alias{rm_empty_vars}
\title{rm empty vars from list}
\usage{
rm_empty_vars(x)
}
\arguments{
\item{x}{list, dataframe, etc.}
}
\value{
list without empty vars
}
\description{
rm empty vars from list
}
