% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_operations.R
\name{rm_na_columns}
\alias{rm_na_columns}
\title{Rm columns with only nas}
\usage{
rm_na_columns(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe with re-ordered columns
}
\description{
Rm columns with only nas
}
\examples{
require(magrittr)
df <- data.frame(
   symbol    = c('A1BG', 'A2M'), 
   id        = c('1',    '2'),
   name      = c('alpha-1-B glycoprotein', 'alpha-2-macroglobulin'), 
   relevance = c(NA_character_, NA_character_),
   version   = c('NA', 'NA'), 
   type      = c('proteincoding', 'proteincoding'))
df \%>\% rm_na_columns()
}
