% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dimnames.R
\name{set_dimnames}
\alias{set_dimnames}
\title{Set dimnames of a matrix}
\usage{
set_dimnames(x, dim_names)
}
\arguments{
\item{x}{matrix}

\item{dim_names}{list}
}
\value{
matrix with set dimnames
}
\description{
Set dimnames of a matrix
}
\examples{
require(magrittr)
x <- matrix(1:4, nrow=2)
dim_names <- list(feature = c('F1', 'F2'), sample = c('S1', 'S2'))
x \%>\% autonomics.support::set_dimnames(dim_names)
}
