% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vextract.R
\name{vextract}
\alias{vextract}
\title{Vectorized extract}
\usage{
vextract(list_of_vectors, extractor_list, simplify = FALSE)
}
\arguments{
\item{list_of_vectors}{list of n vectors}

\item{extractor_list}{list of n integer or character vector}

\item{simplify}{logical: whether to simplify list output}
}
\value{
list of n vectors
}
\description{
Extract vector from list_of_vectors
}
\examples{
require(magrittr)
list_of_vectors <- list(c(L="STD", M="EM15", H="EM30"), 
                        c(L="STD", M="EM15", H="EM30"), 
                        c(L="STD", M="EM15", H="EM30" ))
extractor_list <- list(c('H', 'L'), c('H', 'L'), c('H', 'L'))
list_of_vectors \%>\% autonomics.support::vextract(extractor_list)
list_of_vectors \%>\% autonomics.support::vextract(extractor_list, simplify = TRUE)
}
