% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsprintf.R
\name{vsprintf}
\alias{vsprintf}
\title{Vectorized sprintf}
\usage{
vsprintf(fmt, ..., first_slowest = TRUE)
}
\arguments{
\item{fmt}{format string}

\item{...}{arguments}
}
\value{
character vector
}
\description{
Vectorized sprintf
}
\examples{
vsprintf(fmt = '\%s.\%s', letters[1:3], 1:3)
vsprintf(fmt = '\%s.\%s', letters[1:3], 1:3, first_slowest = FALSE)
}
