% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_dir_with_time_stamp.R
\name{zip_dir}
\alias{zip_dir}
\title{Zip dir}
\usage{
zip_dir(dir, add_time_stamp = TRUE)
}
\arguments{
\item{dir}{directory which needs to be zipped}

\item{add_time_stamp}{whether to add a time stamp (logical)}
}
\description{
Zip dir
}
\examples{
require(magrittr)
my_dir <- tempdir() \%>\% paste0('/test_zip')
dir.create(my_dir)
my_dir \%>\% paste0('/dirA') \%>\%  dir.create()
my_dir \%>\% paste0('/dirB') \%>\%  dir.create()
data.frame(a=1,b=2) \%>\% print2txt(my_dir \%>\% paste0('/dirA/fileA.txt'))
data.frame(a=3,b=4) \%>\% print2txt(my_dir \%>\% paste0('/dirB/fileB.txt'))
zip_dir(my_dir)
}
