% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cfread}
\alias{cfread}
\title{Convenient fread}
\usage{
cfread(file, verbose = FALSE, integer64 = "numeric", data.table = TRUE, ...)
}
\arguments{
\item{file}{passed to \code{\link[data.table]{fread}(input = file)}}

\item{verbose}{FALSE (default) or TRUE. 
Passed to \code{\link[data.table]{fread}(input = file)}}

\item{integer64}{'numeric' (default) or 'integer64': passed to data.table::fread}

\item{data.table}{\code{TRUE} or \code{FALSE}: whether to return as
\code{data.table} (rather than data.frame)}

\item{...}{Passed to \code{\link[data.table]{fread}}.}
}
\description{
Convenient \code{\link[data.table]{fread}} wrapper which sets 
integer64 to 'numeric' to prevent issues with
}
\details{
Convenience wrapper to \code{\link[data.table]{fread}} with preset arguments, 
making the fread operation:
\itemize{
   \item reliable:   integer64 are read as 'numeric'
   \item silent:     verbose set to FALSE and warnings are suppressed
}

The integer64 = 'numeric' specification is required to read integers as 
regular integers rather than 64 bit integers. The problem with 64 bit 
integers is that the numbers display incorrectly when the package 
bit64 is not loaded.
}
