% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonify_strings.R
\name{commonify_strings}
\alias{commonify_strings}
\title{Commonify strings}
\usage{
commonify_strings(x)
}
\arguments{
\item{x}{string}

\item{y}{string}
}
\description{
Commonify strings
}
\examples{
require(magrittr)

# NO DIFFERENCES
x <- c('Retrotransposon Gag-like protein 8B',
       'Retrotransposon Gag-like protein 8B')
x \%>\% commonify_strings()

# TAILS DIFFER
x <- c('Histone H2B type 1-K',
       'Histone H2B type 1-C/E/F/G/I')
x \%>\% commonify_strings()

# MORE COMPLEX DIFFERENCES
x <- c('Fatty acid binding protein, isoform 3',
       'Fatty acid binding protein', 
       'heart-specific Fatty acid binding protein',
       'heart-specific Fatty acid binding protein, isoform 3')
x \%>\% commonify_strings()

}
