% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_outlier.R
\name{is_outlier}
\alias{is_outlier}
\title{Identify outliers in vector}
\usage{
is_outlier(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
logical vector
}
\description{
Identify outliers in vector
}
\examples{
x <- c(1:25,  200)
is_outlier(x)
}
