% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_operations.R
\name{mutate_keeping_rownames}
\alias{mutate_keeping_rownames}
\title{Mutate while preserving rownames}
\usage{
mutate_keeping_rownames(.data, ...)
}
\arguments{
\item{.data}{dataframe}

\item{...}{passed to dplyr::mutate}
}
\value{
data.frame
}
\description{
Mutate while preserving rownames
}
\examples{
.data <- data.frame(organism = c('Homo sapiens', 'Mus musculus'), 
        commonname   = c('human', 'mouse'), row.names = c('Hs', 'Mm'))
.data \%>\% mutate_keeping_rownames(genus = rownames(.))
}
