% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{uniquify}
\alias{uniquify}
\title{Make vector components unique by appending spaces}
\usage{
uniquify(avector)
}
\arguments{
\item{avector}{character or factor vector}
}
\description{
Make vector components unique by appending spaces
}
\seealso{
\code{\link[base]{make.unique}}
}
